requirements.txt
flask
yt-dlp
openai-whisper
flask-limiter
python-dotenv
